github_mvp_generator/requirements.txt
requests==2.31.0
pygithub==2.1.1
python-dotenv==1.0.0
jinja2==3.1.2
openai==1.106.1
groq==0.13.1
flask==2.3.2
flask-cors==4.0.0